Use New Relic also in dev
This removes a dev/prod difference and prevents some possibility of New
Relice-related issues that occur in prod but not in dev. For example if
New Relic were missing from the project's requirements then the command
would crash in production but would not have crashed in dev. Now it'll
crash in dev also.

This shouldn't do anything in dev--since no New Relic envvars are set in
dev, the New Relic agent won't be reporting anything.

This does mean that you *can* report to New Relic from your dev
environment if you want to. For example reporting to a dev or personal
New Relic account in order to test some code that sends custom metrics
to New Relic.